agregado Frm de Sueldo Perfil Histórico

+++ BugTrackingSystem/BugTrackingSystem/Entidades/SueldoPerfilHistorico.cs
@@ -6,7 +6,7 @@ namespace BugTrackingSystem.Entidades
     {
         public Perfil Perfil { get; set; }
         public DateTime Fecha { get; set; }
-        public float Sueldo { get; set; }
+        public decimal Sueldo { get; set; }
         public bool Borrado { get; set; }
     }
 }
